feat: 🔥 book service

+++ dtos/book-dtos.go
@@ -4,11 +4,12 @@ type BookUpdateDTOS struct {
 	ID          uint64 `json:"id" form:"id" binding:"required"`
 	Title       string `json:"title" form:"title" binding:"required"`
 	Description string `json:"description" form:"description" binding:"required"`
-	UserID      uint   `json:"user_id,omitempty" form:"user_id,omitempty"`
+	UserID      uint64 `json:"user_id,omitempty"  form:"user_id,omitempty"`
 }
 
+// BookCreateDTO is is a model that client use when create a new book
 type BookCreateDTOS struct {
 	Title       string `json:"title" form:"title" binding:"required"`
 	Description string `json:"description" form:"description" binding:"required"`
-	UserID      uint   `json:"user_id,omitempty" form:"user_id,omitempty"`
+	UserID      uint64 `json:"user_id,omitempty"  form:"user_id,omitempty"`
 }
